Give the suite its first gate, over the two helpers it kept rewriting
TestSupport.h's mouseEventAt said "Each of those three now calls this."
Sixteen places built a juce::MouseEvent from its fifteen-argument
constructor, in ten files, under four names - eventAt, eventOn, clickAt,
and eight anonymous lambdas. The doc was true when it was written and had
been false for a while, which is the ordinary fate of a claim nothing
checks.

Neither helper was findable. One is a constructor call with no distinctive
name at all; the other, findDescendantWithID, is four lines - too short to
read as duplication - and nine files had it byte-identical but for what
one of them called its first parameter.

The primitive is mouseEventOn, taking both components and both positions,
because two of the sixteen needed what the old helper could not express: a
header button hit while the CARD handles it, and a drag whose mouseDownPos
differs from where it is reported. Passing one point for both makes
getDistanceFromDragStart zero, which is right for a click and silently
wrong for the drag it turns into one. mouseEventAt and mouseDragEvent sit
on top, and mouseEventAt takes a float point only - an int overload beside
it made every braced call ambiguous. Harnesses still wrap it under their
own names where their call sites hold a Point<int>; what they no longer do
is fill the constructor.

findDescendantWithID moves to ControlWalkHarness.h, which is already where
a test goes to walk components, and nine copies go.

The gate needed a new scan: offenders() walks src/ and reports paths
relative to it, so a tests/ path came back as ../tests/Foo.cpp and no
exemption could match. offendersIn takes the files and the root; offenders
is now the src/ wrapper. testFiles() is the tests/ half of allDewFiles.

This file exempts itself, and the reason is worth having in writing: the
scanner copies quoted literals through whole so that a marker a gate
quotes cannot open a comment, so the two needles below are two matches in
the file that spells them. It is the first gate over tests/ and so the
first that could read itself.

Proved by putting a copy of findDescendantWithID back in ConfirmTests: the
gate names the file and the line.

+++ tests/RollHarness.h
@@ -69,7 +69,7 @@ inline juce::MouseEvent eventAt (juce::Component& target, juce::Point<int> local
                                  juce::ModifierKeys mods = juce::ModifierKeys(), int clickCount = 1,
                                  bool wasDragged = false)
 {
-    return mouseEventAt (target, local, mods, clickCount, wasDragged);
+    return mouseEventAt (target, local.toFloat(), mods, clickCount, wasDragged);
 }
 
 inline void clickAndRelease (juce::Component& c, juce::Point<int> at,
